Correct withings id type

+++ sql/health.sql
@@ -24,7 +24,7 @@ create table fitbit_token_gargling (
 
 
 create table withings_token_gargling (
-  withings_id smallint not null references withings_token(id),
+  withings_id int not null references withings_token(id),
   gargling_id smallint not null references gargling(id)
 );
